requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
anthropic>=0.25.0
pdfplumber>=0.10.0
resend>=0.6.0
supabase>=2.0.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
